backend/pyproject.toml
[project]
name = "rental-evaluator-backend"
version = "0.1.0"
description = "Persona-first rental property evaluator — FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.7",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "supabase>=2.28.3",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

# ── How to run ────────────────────────────────────────────────────────────────
# Dev:        uv run uvicorn app.main:app --reload
# Production: uv run uvicorn app.main:app --host 0.0.0.0 --port 8000
